完善

+++ s3cloudcache-common/src/main/java/org/foreverfzl/cloudchache/common/ProjectUtil.java
@@ -223,9 +223,9 @@ public final class ProjectUtil {
      * @param num2 除数，必须是2的整数次幂
      * @return 整除结果
      */
-    public static long divideByPower(long num1, long num2) {
+    public static long divideByPower(long num1, int num2) {
         // 获取2的幂次，即右移位数
-        int shiftBits = Long.numberOfTrailingZeros(num2);
+        int shiftBits = Integer.numberOfTrailingZeros(num2);
         return num1 >> shiftBits;
     }
